[feat] 实现测试多种轨迹 - Minco由于其不保证凸，且过程中有可能导致L-bfgs的近似Hessian病态，不在中层使用 - 多项式曲线仅用于Minco的辅助使用，其曲线特性不适合通过无约束的优化方法进行优化，需要特化实现 - BSpline 实现了均匀B样条通过路点参数化的方法，使用4阶B样条，还未实现优化器

+++ safe_planner/inc/safe_planner/trajectory/trajectroies.hpp
@@ -6,10 +6,10 @@ namespace safe_planner::planner
 {
 class ITrajectory{
 public:
-using seconds = std::chrono::duration<float>;
+using seconds = std::chrono::duration<double>;
 virtual ~ITrajectory() = default;
 
-virtual Eigen::Vector3f pos(const seconds) const = 0;
+virtual Eigen::Vector3d pos(const seconds) const = 0;
 virtual void get_time_range(seconds& from, seconds& to) const = 0;
 };
 }
